refactor(adapter): web 어댑터 패키지 구조 간소화
- adapter.inbound.web -> adapter.web 패키지명 변경
- import 경로 업데이트
