+++ backend/src/main/kotlin/org/setackle/backend/adapter/web/user/dto/RefreshTokenDto.kt
@@ -1,4 +1,4 @@
-package org.setackle.backend.adapter.inbound.web.user.dto
+package org.setackle.backend.adapter.web.user.dto
 
 import jakarta.validation.constraints.NotBlank
 import org.setackle.backend.application.user.inbound.RefreshTokenCommand
